requirements.txt
asgiref==3.11.0
Django==5.2.8
django-filter==25.2
django-tinymce==5.0.0
djangorestframework==3.16.1
PyMySQL==1.1.0
sqlparse==0.5.3
tzdata==2025.2
python-dotenv
gunicorn
django-q
requests